requirements.txt
flask==3.0.0
pyjwt==2.8.0
requests==2.31.0
cryptography==41.0.7